Changed simulation time variables from float to double

+++ MolecularDynamics/BarnesHutParallel/SpaceFillingCurves/include/Integrator.h
@@ -23,7 +23,7 @@
 
 #include "H5Profiler.h"
 
-void timeIntegration_BH_par(float t, float delta_t, float t_end, float diam, TreeNode *root, SubDomainKeyTree *s,
+void timeIntegration_BH_par(double t, double delta_t, double t_end, float diam, TreeNode *root, SubDomainKeyTree *s,
                             Renderer *renderer, char *image, double *hdImage, bool render=true,
                             bool processColoring=false, bool h5Dump=false, int h5DumpEachTimeSteps=1,
                             int loadBalancingInterval=1);
